+++ src/poxy/defaults.py
@@ -591,6 +591,11 @@ class Defaults:
         r'availability': r'@par [parent_parent_set_class m-block m-special]Conditional availability ^^',
         r'figure{1}': r'@image html \1',
         r'figure{2}': r'@image html \1 "\2"',
+        # tabbed content blocks (restructured by the TabBlocks html fixer)
+        r'tabs': r'[tabs]',
+        r'etabs': r'[/tabs]',
+        r'endtabs': r'[/tabs]',
+        r'tab{1}': r'[tab \1]',
         # m.css
         r'm_div{1}': r'@xmlonly<mcss:div xmlns:mcss="http://mcss.mosra.cz/doxygen/" mcss:class="\1">@endxmlonly',
         r'm_enddiv': r'@xmlonly</mcss:div>@endxmlonly',